EditTable refactor
